#1199:  * Change `AbstractTetrahedralMesh::save` so it always writes the in-memory mesh to disk, rather than trying to read an on-disk mesh and write it again.  This ensures we retain any node permutation in the checkpoint.  * Change `CardiacSimulationArchiver::Migrate` so it special-cases if we're loading on the same number of processes, in order to make the next bullet work, by setting up the correct original `DistributedVectorFactory` on every process.  * Change `AbstractTetrahedralMesh::load` so if we're not migrating, it uses the original `DistributedVectorFactory` to specify the node distribution for the mesh.
Also a little tidying.

git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@8260 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ mesh/test/TestNonCachedTetrahedralMesh.hpp
@@ -48,21 +48,28 @@ private:
 #ifdef __linux__
         OutputFileHandler handler("");
         std::string file_name = handler.GetOutputDirectoryFullPath() + "memusage.tmp";
-
-        std::stringstream ps_command;
-        // Option "-o vsize=" makes ps report process virtual size. "=" after "vsize" prints no column header.
-        ps_command << "ps -o vsize= " << getpid() << " > " << file_name;
-        EXPECT0(system, ps_command.str().c_str());
-
+        if (PetscTools::AmMaster())
+        {
+            std::stringstream ps_command;
+            // Option "-o vsize=" makes ps report process virtual size. "=" after "vsize" prints no column header.
+            ps_command << "ps -o vsize= " << getpid() << " > " << file_name;
+            EXPECT0(system, ps_command.str().c_str());
+        }
+        PetscTools::Barrier("GetMemoryUsage-1");
+    
         std::ifstream mem_file;
         mem_file.open(file_name.c_str());
         assert(mem_file.is_open());
         unsigned vsize;
         mem_file >> vsize;
-
-        std::string rm_command = "rm -rf " + file_name;
-        EXPECT0(system, rm_command.c_str());
-
+    
+        PetscTools::Barrier("GetMemoryUsage-2");
+        if (PetscTools::AmMaster())
+        {
+            std::string rm_command = "rm -rf " + file_name;
+            EXPECT0(system, rm_command.c_str());
+        }
+    
         return vsize;
 #else
         return 0;
@@ -98,7 +105,7 @@ public:
         double non_cached_construction_time = (non_cached_finish-non_cached_start)/(double)CLOCKS_PER_SEC;
 
         // Constructing the non cached object should be quicker
-        TS_ASSERT( cached_construction_time > non_cached_construction_time );
+        TS_ASSERT_LESS_THAN( non_cached_construction_time, cached_construction_time );
 
         // compare mem usage
         TS_ASSERT( cached_mem_usage >= non_cached_mem_usage );
